requirements.txt
Flask==2.2.5
openai==1.6.1
python-dotenv==1.0.0
requests==2.31.0
